% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi}
\alias{bmi}
\title{BMI}
\usage{
bmi(x, y)
}
\arguments{
\item{x}{Your size in centimeters.}

\item{y}{Your weight in kilograms.}
}
\description{
Return the value of the body mass index (bmi).
}
\examples{
bmi(x=183, y=75)
}
